Move data list components under page directories
