+++ src/app/home/pages/ClustersPage/index.ts
@@ -1,0 +1,1 @@
+export * from './ClustersPage';
