+++ src/app/home/pages/PlansPage/components/MigrationsTable.module.scss
