+++ src/app/home/pages/PlansPage/components/PlanContent.tsx
