+++ src/app/home/pages/PlansPage/index.ts
@@ -1,0 +1,1 @@
+export * from './PlansPage';
