+++ src/app/home/pages/StoragesPage/components/StorageDataListItem.tsx
