+++ src/app/home/pages/StoragesPage/components/StorageItem.tsx
