+++ src/app/home/pages/StoragesPage/index.ts
@@ -1,0 +1,1 @@
+export * from './StoragesPage';
