+++ src/app/home/pages/index.ts
@@ -1,3 +1,3 @@
 export * from './ClustersPage';
-export * from './PlansPage';
+export * from './PlansPage/PlansPage';
 export * from './StoragesPage';
